refactor(explorer): extract dialogs to file
